Renamed and improved parser methods
