+++ src/AngleSharp/AngleSharp.Core.csproj
@@ -64,6 +64,7 @@
     <Compile Include="Css\Parser\ICssSelectorParser.cs" />
     <Compile Include="Dom\IPseudoElement.cs" />
     <Compile Include="Html\Dom\ImageExtensions.cs" />
+    <Compile Include="Html\Parser\HtmlDomBuilderExtensions.cs" />
     <Compile Include="Html\ValidationErrors.cs" />
     <Compile Include="Html\Parser\IHtmlParser.cs" />
     <Compile Include="Io\BaseRequester.cs" />
@@ -226,6 +227,7 @@
     <Compile Include="Io\ResourceRequest.cs" />
     <Compile Include="Css\Parser\CssParseError.cs" />
     <Compile Include="Html\Parser\HtmlParseError.cs" />
+    <Compile Include="Xml\Parser\XmlParseErrorExtensions.cs" />
     <Compile Include="Xml\Parser\XmlParseException.cs" />
     <Compile Include="Xml\Parser\XmlParser.cs" />
     <Compile Include="Xml\Parser\XmlParserExtensions.cs" />
